Coleta da fase no object selector a partir do game manager
