--- Assets/scripts/Game Manager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 84723951a2734f84b95f5d21cffd170d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
