+++ Assets/scripts/Game_Manager.cs
@@ -1,0 +1,24 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Game_Manager : MonoBehaviour{
+
+    public static int stage = 0;
+    public float time = 0;
+    public float n = 1;
+
+    void Start()
+    {
+        time += Time.deltaTime;
+    }
+
+    void Update()
+    {
+        time += Time.deltaTime;
+        if(time >= 3 * n){
+            stage++;
+            n *= 1.5f;
+        }
+    }
+}
