+++ Assets/scripts/Object_selector.cs
@@ -1,8 +1,8 @@
-using System.Collections;
+﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class ObjectSelector : MonoBehaviour
+public class Object_Selector : MonoBehaviour
 {
 
     public GameObject alien;
@@ -17,26 +17,24 @@ public class ObjectSelector : MonoBehaviour
     public GameObject satelite_1;
     public GameObject satelite_2;
 
-    public GameObject obstacle;
+    public static GameObject obstacle;
 
-    public GameObject[] extras;
     public GameObject[] actual_stage;
 
-    // private int stage_index =  GameManager.stage;
-    private int stage_index = 0;
+    private static int stage_index = Game_Manager.stage;
 
     public int obstacle_index = 0;
     private int stage_length = 0;
 
     void Start()
     {
+        stage_index = Game_Manager.stage;
         GameObject[][] actual_stage = new GameObject[3][];
-
         actual_stage[0] = new GameObject[4] {nuvem_1_0, nuvem_1_1, nuvem_3_0, balao_1};
         actual_stage[1] = new GameObject[3] {meteoro_1_0, meteoro_2_0, meteoro_3_0};
         actual_stage[2] = new GameObject[4] {satelite_1, satelite_2, satelite_3, alien};
         
-        // stage_length = (actual_stage[stage_index]).Count;
+        //stage_length = (actual_stage[stage_index]).Count;
         // obstacle_index = Random.Range(0, stage_length);
 
         GameObject obstacle = actual_stage[stage_index][obstacle_index];
@@ -44,6 +42,8 @@ public class ObjectSelector : MonoBehaviour
 
     void Update()
     {
+        stage_index = Game_Manager.stage;
+        print(stage_index);
 
     }
 }
